perf: switched block id to byte

+++ src/main/java/ovh/paulem/mc/world/block/Blocks.java
@@ -7,21 +7,20 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Blocks {
-    public static final Map<Integer, Block> blocks = new HashMap<>();
+    public static final Map<Byte, Block> blocks = new HashMap<>();
 
-    public static final AirBlock AIR = register(new AirBlock("air", 0));
-    public static final SingleFaceBlock STONE = register(new SingleFaceBlock("stone", 1));
-    public static final SingleFaceBlock DIRT = register(new SingleFaceBlock("dirt", 2));
-    public static final TintTopFaceBlock GRASS_BLOCK = register(new GrassBlock("grass_block", 3, TintType.GRASS));
-    public static final SingleFaceBlock LOG = register(new SingleFaceBlock("log", 4));
-    public static final FoliageBlock LEAVES = register(new FoliageBlock("leaves", 5, true, TintType.FOLIAGE));
+    public static final AirBlock AIR = register(new AirBlock("air", (byte)0));
+    public static final SingleFaceBlock STONE = register(new SingleFaceBlock("stone", (byte)1));
+    public static final SingleFaceBlock DIRT = register(new SingleFaceBlock("dirt", (byte)2));
+    public static final TintTopFaceBlock GRASS_BLOCK = register(new GrassBlock("grass_block", (byte)3, TintType.GRASS));
+    public static final SingleFaceBlock LOG = register(new SingleFaceBlock("log", (byte)4));
+    public static final FoliageBlock LEAVES = register(new FoliageBlock("leaves", (byte)5, true, TintType.FOLIAGE));
 
     public static<T extends Block> T register(T block) {
-        if(blocks.containsKey(block.getId())) {
+        if(blocks.containsKey((byte)block.getId())) {
             throw new IllegalArgumentException("Block with id " + block.getId() + " is already registered!");
         }
-
-        blocks.put(block.getId(), block);
+        blocks.put((byte)block.getId(), block);
         return block;
     }
 }
